Ajuste Dto e Form para listagem e cadastro
