+++ src/main/java/com/alura/easycharge/EasychargeApplication.java
@@ -2,10 +2,12 @@ package com.alura.easycharge;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.data.web.config.EnableSpringDataWebSupport;
 
 @SpringBootApplication
 @EnableSpringDataWebSupport
+@EnableCaching
 public class EasychargeApplication {
 
 	public static void main(String[] args) {
