+++ src/main/java/com/alura/easycharge/config/security/SecurityConfigurations.java
@@ -1,0 +1,33 @@
+package com.alura.easycharge.config.security;
+
+//import org.springframework.context.annotation.Configuration;
+//import org.springframework.http.HttpMethod;
+//import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+//import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+//import org.springframework.security.config.annotation.web.builders.WebSecurity;
+//import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+//import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+//@EnableWebSecurity
+//@Configuration
+//public class SecurityConfigurations extends WebSecurityConfigurerAdapter {
+//
+//
+//    //Configurações de autentificação
+//    @Override
+//    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+//    }
+//
+//    //Configurações de Autorização
+//    @Override
+//    protected void configure(HttpSecurity http) throws Exception {
+//        http.authorizeHttpRequests().antMatchers(HttpMethod.GET, "/api").permitAll();
+//    }
+//
+//    //Configurações de recursos estaticos(js, css, imagens, etc.)
+//    @Override
+//    public void configure(WebSecurity web) throws Exception {
+//    }
+//
+//
+//}
