+++ src/main/java/com/alura/easycharge/dto/ClienteDTO.java
@@ -1,8 +1,10 @@
 package com.alura.easycharge.dto;
 
+import com.alura.easycharge.form.ClienteForm;
 import com.alura.easycharge.models.Cliente;
 import com.alura.easycharge.models.Endereco;
 import com.alura.easycharge.models.StatusCliente;
+import org.springframework.data.domain.Page;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
@@ -58,7 +60,6 @@ public class ClienteDTO {
         this.status = cliente.getStatus();
     }
 
-
     public Long getId() {
         return id;
     }
@@ -67,104 +68,52 @@ public class ClienteDTO {
         return nome;
     }
 
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
     public String getCpf() {
         return cpf;
-    }
-
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-
-    public String getTelefone() {
-        return telefone;
-    }
-
-    public void setTelefone(String telefone) {
-        this.telefone = telefone;
     }
 
     public String getEmail() {
         return email;
     }
 
-    public void setEmail(String email) {
-        this.email = email;
+    public String getTelefone() {
+        return telefone;
     }
 
     public String getRua() {
         return rua;
     }
 
-    public void setRua(String rua) {
-        this.rua = rua;
-    }
-
     public String getNumero() {
         return numero;
-    }
-
-    public void setNumero(String numero) {
-        this.numero = numero;
     }
 
     public String getComplemento() {
         return complemento;
     }
 
-    public void setComplemento(String complemento) {
-        this.complemento = complemento;
-    }
-
     public String getBairro() {
         return bairro;
-    }
-
-    public void setBairro(String bairro) {
-        this.bairro = bairro;
     }
 
     public String getCidade() {
         return cidade;
     }
 
-    public void setCidade(String cidade) {
-        this.cidade = cidade;
-    }
-
     public String getEstado() {
         return estado;
-    }
-
-    public void setEstado(String estado) {
-        this.estado = estado;
     }
 
     public String getProfissao() {
         return profissao;
     }
 
-    public void setProfissao(String profissao) {
-        this.profissao = profissao;
-    }
-
     public BigDecimal getRenda() {
         return renda;
     }
 
-    public void setRenda(BigDecimal renda) {
-        this.renda = renda;
-    }
-
     public StatusCliente getStatus() {
         return status;
-    }
-
-    public void setStatus(StatusCliente status) {
-        this.status = status;
     }
 
     public Cliente toCliente() {
@@ -181,4 +130,7 @@ public class ClienteDTO {
         return cliente;
     }
 
+    public static Page<ClienteDTO> converter(Page<Cliente> clientes){
+        return clientes.map(ClienteDTO::new);
+    }
 }
