+++ src/main/resources/application.properties
@@ -2,5 +2,6 @@ spring.jpa.hibernate.ddl-auto=update
 spring.datasource.url=jdbc:mysql://localhost:3306/easycharge?useTimezone=true&serverTimezone=UTC
 spring.datasource.username=root
 spring.datasource.password=
-
+spring.jpa.properties.hibernate.show_sql=true
+spring.jpa.properties.hibernate.format_sql=true
